+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityEditPanel.java
@@ -259,10 +259,11 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 	}
 
 	/**
-	 * Performs insert on the active entity after asking for confirmation using the {@link Confirmer}
+	 * <p>Performs insert on the active entity after asking for confirmation using the {@link Confirmer}
 	 * specified via {@link Config#insertConfirmer(Confirmer)}.
-	 * Note that the default insert {@link Confirmer} simply returns true, so in order to implement
+	 * <p>Note that the default insert {@link Confirmer} simply returns true, so in order to implement
 	 * an insert confirmation you must set the {@link Confirmer} via {@link Config#insertConfirmer(Confirmer)}.
+	 * <p>This method executes synchronously on the EDT.
 	 * @return true in case of successful insert, false otherwise
 	 * @see Config#insertConfirmer(Confirmer)
 	 */
@@ -275,7 +276,8 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 	}
 
 	/**
-	 * Performs insert on the active entity without asking for confirmation
+	 * <p>Performs insert on the active entity without asking for confirmation.
+	 * <p>This method executes synchronously on the EDT.
 	 * @return true in case of successful insert, false otherwise
 	 */
 	public final boolean insert() {
@@ -290,21 +292,17 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 
 			return true;
 		}
-		catch (ValidationException e) {
-			LOG.debug(e.getMessage(), e);
-			onException(e);
-		}
-		catch (Exception e) {
-			LOG.error(e.getMessage(), e);
-			onException(e);
+		catch (Exception exception) {
+			handleException(exception);
 		}
 
 		return false;
 	}
 
 	/**
-	 * Performs delete on the active entity after asking for confirmation using the {@link Confirmer}
+	 * <p>Performs delete on the active entity after asking for confirmation using the {@link Confirmer}
 	 * specified via {@link Config#deleteConfirmer(Confirmer)}.
+	 * <p>This method executes synchronously on the EDT.
 	 * @return true if the delete operation was successful
 	 * @see Config#deleteConfirmer(Confirmer)
 	 */
@@ -317,7 +315,8 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 	}
 
 	/**
-	 * Performs delete on the active entity without asking for confirmation
+	 * <p>Performs delete on the active entity without asking for confirmation
+	 * <p>This method executes synchronously on the EDT.
 	 * @return true if the delete operation was successful
 	 */
 	public final boolean delete() {
@@ -327,21 +326,17 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 
 			return true;
 		}
-		catch (ReferentialIntegrityException e) {
-			LOG.debug(e.getMessage(), e);
-			onException(e);
-		}
-		catch (Exception ex) {
-			LOG.error(ex.getMessage(), ex);
-			onException(ex);
+		catch (Exception exception) {
+			handleException(exception);
 		}
 
 		return false;
 	}
 
 	/**
-	 * Performs update on the active entity after asking for confirmation using the {@link Confirmer}
+	 * <p>Performs update on the active entity after asking for confirmation using the {@link Confirmer}
 	 * specified via {@link Config#updateConfirmer(Confirmer)}.
+	 * <p>This method executes synchronously on the EDT.
 	 * @return true if the update operation was successful
 	 * @see Config#updateConfirmer(Confirmer)
 	 */
@@ -354,7 +349,8 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 	}
 
 	/**
-	 * Performs update on the active entity without asking for confirmation.
+	 * <p>Performs update on the active entity without asking for confirmation.
+	 * <p>This method executes synchronously on the EDT.
 	 * @return true if the update operation was successful
 	 */
 	public final boolean update() {
@@ -364,13 +360,8 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 
 			return true;
 		}
-		catch (ValidationException e) {
-			LOG.debug(e.getMessage(), e);
-			onException(e);
-		}
-		catch (Exception ex) {
-			LOG.error(ex.getMessage(), ex);
-			onException(ex);
+		catch (Exception exception) {
+			handleException(exception);
 		}
 
 		return false;
@@ -732,6 +723,16 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 		return new Config(config);
 	}
 
+	private void handleException(Exception exception) {
+		if (exception instanceof ValidationException || exception instanceof ReferentialIntegrityException) {
+			LOG.debug(exception.getMessage(), exception);
+		}
+		else {
+			LOG.error(exception.getMessage(), exception);
+		}
+		onException(exception);
+	}
+
 	private static boolean componentSelectable(JComponent component) {
 		return component != null &&
 						component.isDisplayable() &&
@@ -1074,7 +1075,7 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 								.title(MESSAGES.getString("inserting"))
 								.owner(EntityEditPanel.this)
 								.onResult(this::handleResult)
-								.onException(this::onException)
+								.onException(EntityEditPanel.this::handleException)
 								.execute();
 			}
 		}
@@ -1087,11 +1088,6 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 			if (configuration.requestFocusAfterInsert) {
 				focus().afterInsert().request();
 			}
-		}
-
-		private void onException(Exception exception) {
-			LOG.error(exception.getMessage(), exception);
-			EntityEditPanel.this.onException(exception);
 		}
 	}
 
@@ -1112,7 +1108,7 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 								.title(MESSAGES.getString("updating"))
 								.owner(EntityEditPanel.this)
 								.onResult(this::handleResult)
-								.onException(this::onException)
+								.onException(EntityEditPanel.this::handleException)
 								.execute();
 			}
 		}
@@ -1120,11 +1116,6 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 		private void handleResult(UpdateEntities.Result result) {
 			onUpdate.accept(result.handle());
 			focus().afterUpdate().request();
-		}
-
-		private void onException(Exception exception) {
-			LOG.error(exception.getMessage(), exception);
-			EntityEditPanel.this.onException(exception);
 		}
 	}
 
@@ -1145,7 +1136,7 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 								.title(MESSAGES.getString("deleting"))
 								.owner(EntityEditPanel.this)
 								.onResult(this::handleResult)
-								.onException(this::onException)
+								.onException(EntityEditPanel.this::handleException)
 								.execute();
 			}
 		}
@@ -1153,11 +1144,6 @@ public abstract class EntityEditPanel extends EntityEditComponentPanel {
 		private void handleResult(DeleteEntities.Result result) {
 			onDelete.accept(result.handle());
 			focus().initial().request();
-		}
-
-		private void onException(Exception exception) {
-			LOG.error(exception.getMessage(), exception);
-			EntityEditPanel.this.onException(exception);
 		}
 	}