+++ changelog.md
@@ -13,6 +13,8 @@ Codion Change Log
 ### is.codion.swing.framework.model
 - DefaultEntityComboBoxModel.DefaultForeignKeyFilter.link() bug fixed, combo box model used its own selection instead of the one from the filter model when initializing the link
 - SwingEntityEditModel.configureComboBoxModel() added, for separating the configuration of a combo box model for use in the edit model from combo box model creation.
+### is.codion.swing.framework.ui
+- EntityEditPanel, exception handling and logging now consistent for insert, update and delete.
 
 ## 0.18.32
 ### is.codion.common.model